Safari-611.1.10/LayoutTests/fast/dom/Geolocation/timeout-expected.txt

Tests that when timeout is non-zero, the success callback is called as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS position.coords.latitude is mockLatitude
PASS position.coords.longitude is mockLongitude
PASS position.coords.accuracy is mockAccuracy
PASS successfullyParsed is true

TEST COMPLETE
